Move ValueOrChunk from execution_engine out to casper_node::types
