+++ node/src/types.rs
@@ -15,6 +15,7 @@ mod node_id;
 pub mod peers_map;
 mod shared_object;
 mod status_feed;
+mod value_or_chunk;
 
 use rand::{CryptoRng, RngCore};
 #[cfg(not(test))]
@@ -46,6 +47,7 @@ pub use node_config::NodeConfig;
 pub(crate) use node_id::NodeId;
 pub use peers_map::PeersMap;
 pub use status_feed::{ChainspecInfo, GetStatusResult, NodeState, StatusFeed};
+pub use value_or_chunk::*;
 
 /// An object-safe RNG trait that requires a cryptographically strong random number generator.
 pub trait CryptoRngCore: CryptoRng + RngCore {}
